feat: search 관련 interface 폴더로 분리 - 여러 파일에서 공동으로 사용하는 interface 분리
