+++ src/search/interface/Answer.tsx
@@ -1,0 +1,10 @@
+interface Memo {
+  id: string;
+  content: string;
+  tags: string[];
+}
+
+export interface Answer {
+  text: string;
+  memos: Memo[] | undefined;
+}
