+++ src/search/interface/SearchQuery.tsx
@@ -1,0 +1,7 @@
+import { Answer } from "./Answer";
+
+export interface SearchQuery {
+  id: string;
+  query: string;
+  answer: Answer;
+}
